[ch #155710913] fix tests structure
