+++ src/__tests__/common-components/private-routes.test.js
@@ -1,12 +1,10 @@
 import React from 'react';
 import { shallow, mount, render } from 'enzyme';
-import Footer from '../components/common/footer';
+import Private from '../../containers/private';
 import toJson, { shallowToJson } from 'enzyme-to-json';
 
-describe('Footer component', () => {
-    const footer = jest.fn();
-    const wrapper = shallow(<Footer footer={footer} location={{}} />);
-    const preventDefault = jest.fn();
+describe('Login component', () => {
+    const wrapper = shallow(<Private />);
 
     it('renders properly without crashing', () => {
         expect(shallowToJson(wrapper)).toMatchSnapshot();
